+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 History of user-visible changes.
 
 ## [v1.5.0](https://github.com/openconnect/openconnect-gui/tree/v1.5.0) (unreleased)
+- New simple profile setup&connect dialog
 - Simple MacOS installer (#81)
 - update build scripts to Qt 5.7
 - update to OpenConnect v7.08